+++ addon/globalPlugins/polyglot/common/config.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 
 import config as nvdaConfig
+import extensionPoints
 from typing import Any
 
 CONF_SECTION = "modernTranslate"
+localDictionarySettingsChanged = extensionPoints.Action()
 
 
 def getConfig() -> dict[str, Any]: