+++ addon/globalPlugins/polyglot/common/languages.py
@@ -20,6 +20,7 @@ ALL_LANGUAGES = {
 	"zh-CHS": _("Chinese (Simplified)"),
 	"zh-Hans": _("Chinese (Simplified)"),
 	"zh-TW": _("Chinese (Traditional)"),
+	"zh-HK": _("Chinese (Traditional)"),
 	"zh-Hant": _("Chinese (Traditional)"),
 	"zh_HANT": _("Chinese (Traditional)"),  # Lingva
 	"cht": _("Chinese (Traditional)"),